Split common test functions to common source file
